move Shuffle layers to layers.py completely
