+++ dask/dataframe/shuffle.py
@@ -1,9 +1,7 @@
 import contextlib
-from collections import defaultdict
 import logging
 import math
 import shutil
-import operator
 import uuid
 import tempfile
 
@@ -11,359 +9,18 @@ import tlz as toolz
 import numpy as np
 import pandas as pd
 
-from .core import DataFrame, Series, _Frame, _concat, map_partitions, new_dd_object
+from .core import DataFrame, Series, _Frame, map_partitions, new_dd_object
 
 from .. import base, config
 from ..base import tokenize, compute, compute_as_if_collection, is_dask_collection
 from ..highlevelgraph import HighLevelGraph
 from ..sizeof import sizeof
-from ..utils import digit, insert, M
+from ..utils import digit, M
 from .utils import hash_object_dispatch, group_split_dispatch
 from . import methods
-from ..layers import SimpleShuffleGraph, ShuffleGraph
+from ..layers import SimpleShuffleLayer, ShuffleLayer
 
 logger = logging.getLogger(__name__)
-
-
-class SimpleShuffleLayer(SimpleShuffleGraph):
-    """Simple HighLevelGraph Shuffle layer
-
-    High-level graph layer for a simple shuffle operation in which
-    each output partition depends on all input partitions.
-
-    Parameters
-    ----------
-    name : str
-        Name of new shuffled output collection.
-    column : str or list of str
-        Column(s) to be used to map rows to output partitions (by hashing).
-    npartitions : int
-        Number of output partitions.
-    npartitions_input : int
-        Number of partitions in the original (un-shuffled) DataFrame.
-    ignore_index: bool, default False
-        Ignore index during shuffle.  If ``True``, performance may improve,
-        but index values will not be preserved.
-    name_input : str
-        Name of input collection.
-    meta_input : pd.DataFrame-like object
-        Empty metadata of input collection.
-    parts_out : list of int (optional)
-        List of required output-partition indices.
-    annotations : dict (optional)
-        Layer annotations
-    """
-
-    def __init__(
-        self,
-        name,
-        column,
-        npartitions,
-        npartitions_input,
-        ignore_index,
-        name_input,
-        meta_input,
-        parts_out=None,
-        annotations=None,
-    ):
-        super().__init__(annotations=annotations)
-        self.name = name
-        self.column = column
-        self.npartitions = npartitions
-        self.npartitions_input = npartitions_input
-        self.ignore_index = ignore_index
-        self.name_input = name_input
-        self.meta_input = meta_input
-        self.parts_out = parts_out or range(npartitions)
-
-    def get_output_keys(self):
-        return {(self.name, part) for part in self.parts_out}
-
-    def __repr__(self):
-        return "SimpleShuffleLayer<name='{}', npartitions={}>".format(
-            self.name, self.npartitions
-        )
-
-    def is_materialized(self):
-        return hasattr(self, "_cached_dict")
-
-    @property
-    def _dict(self):
-        """Materialize full dict representation"""
-        if hasattr(self, "_cached_dict"):
-            return self._cached_dict
-        else:
-            dsk = self._construct_graph()
-            self._cached_dict = dsk
-        return self._cached_dict
-
-    def __getitem__(self, key):
-        return self._dict[key]
-
-    def __iter__(self):
-        return iter(self._dict)
-
-    def __len__(self):
-        return len(self._dict)
-
-    def __reduce__(self):
-        attrs = [
-            "name",
-            "column",
-            "npartitions",
-            "npartitions_input",
-            "ignore_index",
-            "name_input",
-            "meta_input",
-            "parts_out",
-            "annotations",
-        ]
-        return (SimpleShuffleLayer, tuple(getattr(self, attr) for attr in attrs))
-
-    def __dask_distributed_pack__(
-        self, all_hlg_keys, known_key_dependencies, client, client_keys
-    ):
-        from distributed.protocol.serialize import to_serialize
-
-        return {
-            "name": self.name,
-            "column": self.column,
-            "npartitions": self.npartitions,
-            "npartitions_input": self.npartitions_input,
-            "ignore_index": self.ignore_index,
-            "name_input": self.name_input,
-            "meta_input": to_serialize(self.meta_input),
-            "parts_out": list(self.parts_out),
-            "concat_func": {"__module__": _concat.__module__, "__name__": "_concat"},
-            "getitem_func": {"__module__": "operator", "__name__": "getitem"},
-            "shuffle_group_func": {
-                "__module__": shuffle_group.__module__,
-                "__name__": "shuffle_group",
-            },
-        }
-
-    def _keys_to_parts(self, keys):
-        """Simple utility to convert keys to partition indices."""
-        parts = set()
-        for key in keys:
-            try:
-                _name, _part = key
-            except ValueError:
-                continue
-            if _name != self.name:
-                continue
-            parts.add(_part)
-        return parts
-
-    def _cull_dependencies(self, keys, parts_out=None):
-        """Determine the necessary dependencies to produce `keys`.
-
-        For a simple shuffle, output partitions always depend on
-        all input partitions. This method does not require graph
-        materialization.
-        """
-        deps = defaultdict(set)
-        parts_out = parts_out or self._keys_to_parts(keys)
-        for part in parts_out:
-            deps[(self.name, part)] |= {
-                (self.name_input, i) for i in range(self.npartitions_input)
-            }
-        return deps
-
-    def _cull(self, parts_out):
-        return SimpleShuffleLayer(
-            self.name,
-            self.column,
-            self.npartitions,
-            self.npartitions_input,
-            self.ignore_index,
-            self.name_input,
-            self.meta_input,
-            parts_out=parts_out,
-        )
-
-    def cull(self, keys, all_keys):
-        """Cull a SimpleShuffleLayer HighLevelGraph layer.
-
-        The underlying graph will only include the necessary
-        tasks to produce the keys (indicies) included in `parts_out`.
-        Therefore, "culling" the layer only requires us to reset this
-        parameter.
-        """
-        parts_out = self._keys_to_parts(keys)
-        culled_deps = self._cull_dependencies(keys, parts_out=parts_out)
-        if parts_out != self.parts_out:
-            culled_layer = self._cull(parts_out)
-            return culled_layer, culled_deps
-        else:
-            return self, culled_deps
-
-    def _construct_graph(self):
-        """Construct graph for a simple shuffle operation."""
-        return SimpleShuffleGraph._construct_graph(
-            name=self.name,
-            column=self.column,
-            npartitions=self.npartitions,
-            npartitions_input=self.npartitions_input,
-            ignore_index=self.ignore_index,
-            name_input=self.name_input,
-            meta_input=self.meta_input,
-            parts_out=self.parts_out,
-            concat_func=_concat,
-            getitem_func=operator.getitem,
-            shuffle_group_func=shuffle_group,
-        )
-
-
-class ShuffleLayer(SimpleShuffleLayer, ShuffleGraph):
-    """Shuffle-stage HighLevelGraph layer
-
-    High-level graph layer corresponding to a single stage of
-    a multi-stage inter-partition shuffle operation.
-
-    Stage: (shuffle-group) -> (shuffle-split) -> (shuffle-join)
-
-    Parameters
-    ----------
-    name : str
-        Name of new (partially) shuffled collection.
-    column : str or list of str
-        Column(s) to be used to map rows to output partitions (by hashing).
-    inputs : list of tuples
-        Each tuple dictates the data movement for a specific partition.
-    stage : int
-        Index of the current shuffle stage.
-    npartitions : int
-        Number of output partitions for the full (multi-stage) shuffle.
-    npartitions_input : int
-        Number of partitions in the original (un-shuffled) DataFrame.
-    k : int
-        A partition is split into this many groups during each stage.
-    ignore_index: bool, default False
-        Ignore index during shuffle.  If ``True``, performance may improve,
-        but index values will not be preserved.
-    name_input : str
-        Name of input collection.
-    meta_input : pd.DataFrame-like object
-        Empty metadata of input collection.
-    parts_out : list of int (optional)
-        List of required output-partition indices.
-    annotations : dict (optional)
-        Layer annotations
-    """
-
-    def __init__(
-        self,
-        name,
-        column,
-        inputs,
-        stage,
-        npartitions,
-        npartitions_input,
-        nsplits,
-        ignore_index,
-        name_input,
-        meta_input,
-        parts_out=None,
-        annotations=None,
-    ):
-        super().__init__(
-            name,
-            column,
-            npartitions,
-            npartitions_input,
-            ignore_index,
-            name_input,
-            meta_input,
-            parts_out=parts_out or range(len(inputs)),
-            annotations=annotations,
-        )
-        self.inputs = inputs
-        self.stage = stage
-        self.nsplits = nsplits
-
-    def __repr__(self):
-        return "ShuffleLayer<name='{}', stage={}, nsplits={}, npartitions={}>".format(
-            self.name, self.stage, self.nsplits, self.npartitions
-        )
-
-    def __reduce__(self):
-        attrs = [
-            "name",
-            "column",
-            "inputs",
-            "stage",
-            "npartitions",
-            "npartitions_input",
-            "nsplits",
-            "ignore_index",
-            "name_input",
-            "meta_input",
-            "parts_out",
-            "annotations",
-        ]
-
-        return (ShuffleLayer, tuple(getattr(self, attr) for attr in attrs))
-
-    def __dask_distributed_pack__(self, *args, **kwargs):
-        ret = super().__dask_distributed_pack__(*args, **kwargs)
-        ret["inputs"] = self.inputs
-        ret["stage"] = self.stage
-        ret["nsplits"] = self.nsplits
-        return ret
-
-    def _cull_dependencies(self, keys, parts_out=None):
-        """Determine the necessary dependencies to produce `keys`.
-
-        Does not require graph materialization.
-        """
-        deps = defaultdict(set)
-        parts_out = parts_out or self._keys_to_parts(keys)
-        inp_part_map = {inp: i for i, inp in enumerate(self.inputs)}
-        for part in parts_out:
-            out = self.inputs[part]
-            for k in range(self.nsplits):
-                _inp = insert(out, self.stage, k)
-                _part = inp_part_map[_inp]
-                if self.stage == 0 and _part >= self.npartitions_input:
-                    deps[(self.name, part)].add(("group-" + self.name, _inp, "empty"))
-                else:
-                    deps[(self.name, part)].add((self.name_input, _part))
-        return deps
-
-    def _cull(self, parts_out):
-        return ShuffleLayer(
-            self.name,
-            self.column,
-            self.inputs,
-            self.stage,
-            self.npartitions,
-            self.npartitions_input,
-            self.nsplits,
-            self.ignore_index,
-            self.name_input,
-            self.meta_input,
-            parts_out=parts_out,
-        )
-
-    def _construct_graph(self):
-        """Construct graph for a "rearrange-by-column" stage."""
-        return ShuffleGraph._construct_graph(
-            name=self.name,
-            column=self.column,
-            npartitions=self.npartitions,
-            npartitions_input=self.npartitions_input,
-            ignore_index=self.ignore_index,
-            name_input=self.name_input,
-            meta_input=self.meta_input,
-            parts_out=self.parts_out,
-            concat_func=_concat,
-            getitem_func=operator.getitem,
-            shuffle_group_func=shuffle_group,
-            inputs=self.inputs,
-            stage=self.stage,
-            nsplits=self.nsplits,
-        )
 
 
 def set_index(
