+++ dask/layers.py
@@ -1,4 +1,6 @@
 import tlz as toolz
+import operator
+from collections import defaultdict
 from importlib import import_module
 
 from .core import keys_in_tasks
@@ -6,25 +8,190 @@ from .utils import insert, stringify, stringify_collection_keys
 from .highlevelgraph import Layer
 
 
-def run_ext_function(func, *args):
-    if isinstance(func, dict):
-        func = getattr(import_module(func["__module__"]), func["__name__"])
-    return func(*args)
+class CallableLazyImport:
+    """Function Wrapper for Lazy Importing"""
+
+    def __init__(self, module, name):
+        self.module = module
+        self.name = name
+
+    def __call__(self, *args):
+        func = getattr(import_module(self.module), self.name)
+        return func(*args)
 
 
-class SimpleShuffleGraph(Layer):
-    """Layer-materialization class for shuffle.SimpleShuffleLayer"""
+class SimpleShuffleLayer(Layer):
+    """Simple HighLevelGraph Shuffle layer
 
-    def __init__(self, annotations=None):
+    High-level graph layer for a simple shuffle operation in which
+    each output partition depends on all input partitions.
+
+    Parameters
+    ----------
+    name : str
+        Name of new shuffled output collection.
+    column : str or list of str
+        Column(s) to be used to map rows to output partitions (by hashing).
+    npartitions : int
+        Number of output partitions.
+    npartitions_input : int
+        Number of partitions in the original (un-shuffled) DataFrame.
+    ignore_index: bool, default False
+        Ignore index during shuffle.  If ``True``, performance may improve,
+        but index values will not be preserved.
+    name_input : str
+        Name of input collection.
+    meta_input : pd.DataFrame-like object
+        Empty metadata of input collection.
+    parts_out : list of int (optional)
+        List of required output-partition indices.
+    annotations : dict (optional)
+        Layer annotations
+    """
+
+    def __init__(
+        self,
+        name,
+        column,
+        npartitions,
+        npartitions_input,
+        ignore_index,
+        name_input,
+        meta_input,
+        parts_out=None,
+        annotations=None,
+    ):
         super().__init__(annotations=annotations)
+        self.name = name
+        self.column = column
+        self.npartitions = npartitions
+        self.npartitions_input = npartitions_input
+        self.ignore_index = ignore_index
+        self.name_input = name_input
+        self.meta_input = meta_input
+        self.parts_out = parts_out or range(npartitions)
+        self.concat_func = CallableLazyImport("dask.dataframe.core", "_concat")
+        self.shuffle_group_func = CallableLazyImport(
+            "dask.dataframe.shuffle", "shuffle_group"
+        )
+
+    def get_output_keys(self):
+        return {(self.name, part) for part in self.parts_out}
+
+    def __repr__(self):
+        return "SimpleShuffleLayer<name='{}', npartitions={}>".format(
+            self.name, self.npartitions
+        )
+
+    def is_materialized(self):
+        return hasattr(self, "_cached_dict")
 
     @property
-    def layer_materialize_module(self):
-        return SimpleShuffleGraph.__module__
+    def _dict(self):
+        """Materialize full dict representation"""
+        if hasattr(self, "_cached_dict"):
+            return self._cached_dict
+        else:
+            dsk = self._construct_graph()
+            self._cached_dict = dsk
+        return self._cached_dict
 
-    @property
-    def layer_materialize_class(self):
-        return "SimpleShuffleGraph"
+    def __getitem__(self, key):
+        return self._dict[key]
+
+    def __iter__(self):
+        return iter(self._dict)
+
+    def __len__(self):
+        return len(self._dict)
+
+    def _keys_to_parts(self, keys):
+        """Simple utility to convert keys to partition indices."""
+        parts = set()
+        for key in keys:
+            try:
+                _name, _part = key
+            except ValueError:
+                continue
+            if _name != self.name:
+                continue
+            parts.add(_part)
+        return parts
+
+    def _cull_dependencies(self, keys, parts_out=None):
+        """Determine the necessary dependencies to produce `keys`.
+
+        For a simple shuffle, output partitions always depend on
+        all input partitions. This method does not require graph
+        materialization.
+        """
+        deps = defaultdict(set)
+        parts_out = parts_out or self._keys_to_parts(keys)
+        for part in parts_out:
+            deps[(self.name, part)] |= {
+                (self.name_input, i) for i in range(self.npartitions_input)
+            }
+        return deps
+
+    def _cull(self, parts_out):
+        return SimpleShuffleLayer(
+            self.name,
+            self.column,
+            self.npartitions,
+            self.npartitions_input,
+            self.ignore_index,
+            self.name_input,
+            self.meta_input,
+            parts_out=parts_out,
+        )
+
+    def cull(self, keys, all_keys):
+        """Cull a SimpleShuffleLayer HighLevelGraph layer.
+
+        The underlying graph will only include the necessary
+        tasks to produce the keys (indicies) included in `parts_out`.
+        Therefore, "culling" the layer only requires us to reset this
+        parameter.
+        """
+        parts_out = self._keys_to_parts(keys)
+        culled_deps = self._cull_dependencies(keys, parts_out=parts_out)
+        if parts_out != self.parts_out:
+            culled_layer = self._cull(parts_out)
+            return culled_layer, culled_deps
+        else:
+            return self, culled_deps
+
+    def __reduce__(self):
+        attrs = [
+            "name",
+            "column",
+            "npartitions",
+            "npartitions_input",
+            "ignore_index",
+            "name_input",
+            "meta_input",
+            "parts_out",
+            "annotations",
+        ]
+        return (SimpleShuffleLayer, tuple(getattr(self, attr) for attr in attrs))
+
+    def __dask_distributed_pack__(
+        self, all_hlg_keys, known_key_dependencies, client, client_keys
+    ):
+        from distributed.protocol.serialize import to_serialize
+
+        return {
+            "name": self.name,
+            "column": self.column,
+            "npartitions": self.npartitions,
+            "npartitions_input": self.npartitions_input,
+            "ignore_index": self.ignore_index,
+            "name_input": self.name_input,
+            "meta_input": to_serialize(self.meta_input),
+            "parts_out": list(self.parts_out),
+            "concat_func": self.concat_func,
+            "shuffle_group_func": self.shuffle_group_func,
+        }
 
     @classmethod
     def __dask_distributed_unpack__(cls, state, dsk, dependencies):
@@ -38,7 +205,7 @@ class SimpleShuffleGraph(Layer):
             state["inputs"] = list(state["inputs"])
 
         # Materialize the layer
-        layer_dsk = cls._construct_graph(**state)
+        layer_dsk = cls(**state)._construct_graph()
 
         # Convert all keys to strings and dump tasks
         layer_dsk = {
@@ -51,114 +218,204 @@ class SimpleShuffleGraph(Layer):
 
         return {"dsk": toolz.valmap(dumps_task, layer_dsk), "deps": deps}
 
-    @staticmethod
-    def _construct_graph(
-        name=None,
-        column=None,
-        npartitions=None,
-        npartitions_input=None,
-        ignore_index=None,
-        name_input=None,
-        meta_input=None,
-        parts_out=None,
-        concat_func=None,
-        getitem_func=None,
-        shuffle_group_func=None,
-    ):
+    def _construct_graph(self):
         """Construct graph for a simple shuffle operation."""
 
-        shuffle_group_name = "group-" + name
-        shuffle_split_name = "split-" + name
+        shuffle_group_name = "group-" + self.name
+        shuffle_split_name = "split-" + self.name
 
         dsk = {}
-        for part_out in parts_out:
+        for part_out in self.parts_out:
             _concat_list = [
                 (shuffle_split_name, part_out, part_in)
-                for part_in in range(npartitions_input)
+                for part_in in range(self.npartitions_input)
             ]
-            dsk[(name, part_out)] = (
-                run_ext_function,
-                concat_func,
+            dsk[(self.name, part_out)] = (
+                self.concat_func,
                 _concat_list,
-                ignore_index,
+                self.ignore_index,
             )
             for _, _part_out, _part_in in _concat_list:
                 dsk[(shuffle_split_name, _part_out, _part_in)] = (
-                    run_ext_function,
-                    getitem_func,
+                    operator.getitem,
                     (shuffle_group_name, _part_in),
                     _part_out,
                 )
                 if (shuffle_group_name, _part_in) not in dsk:
                     dsk[(shuffle_group_name, _part_in)] = (
-                        run_ext_function,
-                        shuffle_group_func,
-                        (name_input, _part_in),
-                        column,
+                        self.shuffle_group_func,
+                        (self.name_input, _part_in),
+                        self.column,
                         0,
-                        npartitions,
-                        npartitions,
-                        ignore_index,
-                        npartitions,
+                        self.npartitions,
+                        self.npartitions,
+                        self.ignore_index,
+                        self.npartitions,
                     )
 
         return dsk
 
 
-class ShuffleGraph(SimpleShuffleGraph):
-    """Layer-materialization class for shuffle.ShuffleLayer"""
+class ShuffleLayer(SimpleShuffleLayer):
+    """Shuffle-stage HighLevelGraph layer
 
-    @property
-    def layer_materialize_class(self):
-        return "ShuffleGraph"
+    High-level graph layer corresponding to a single stage of
+    a multi-stage inter-partition shuffle operation.
 
-    @staticmethod
-    def _construct_graph(
-        name=None,
-        column=None,
-        npartitions=None,
-        npartitions_input=None,
-        ignore_index=None,
-        name_input=None,
-        meta_input=None,
+    Stage: (shuffle-group) -> (shuffle-split) -> (shuffle-join)
+
+    Parameters
+    ----------
+    name : str
+        Name of new (partially) shuffled collection.
+    column : str or list of str
+        Column(s) to be used to map rows to output partitions (by hashing).
+    inputs : list of tuples
+        Each tuple dictates the data movement for a specific partition.
+    stage : int
+        Index of the current shuffle stage.
+    npartitions : int
+        Number of output partitions for the full (multi-stage) shuffle.
+    npartitions_input : int
+        Number of partitions in the original (un-shuffled) DataFrame.
+    k : int
+        A partition is split into this many groups during each stage.
+    ignore_index: bool, default False
+        Ignore index during shuffle.  If ``True``, performance may improve,
+        but index values will not be preserved.
+    name_input : str
+        Name of input collection.
+    meta_input : pd.DataFrame-like object
+        Empty metadata of input collection.
+    parts_out : list of int (optional)
+        List of required output-partition indices.
+    annotations : dict (optional)
+        Layer annotations
+    """
+
+    def __init__(
+        self,
+        name,
+        column,
+        inputs,
+        stage,
+        npartitions,
+        npartitions_input,
+        nsplits,
+        ignore_index,
+        name_input,
+        meta_input,
         parts_out=None,
-        concat_func=None,
-        getitem_func=None,
-        shuffle_group_func=None,
-        inputs=None,
-        stage=None,
-        nsplits=None,
+        annotations=None,
     ):
+        super().__init__(
+            name,
+            column,
+            npartitions,
+            npartitions_input,
+            ignore_index,
+            name_input,
+            meta_input,
+            parts_out=parts_out or range(len(inputs)),
+            annotations=annotations,
+        )
+        self.inputs = inputs
+        self.stage = stage
+        self.nsplits = nsplits
+
+    def __repr__(self):
+        return "ShuffleLayer<name='{}', stage={}, nsplits={}, npartitions={}>".format(
+            self.name, self.stage, self.nsplits, self.npartitions
+        )
+
+    def __reduce__(self):
+        attrs = [
+            "name",
+            "column",
+            "inputs",
+            "stage",
+            "npartitions",
+            "npartitions_input",
+            "nsplits",
+            "ignore_index",
+            "name_input",
+            "meta_input",
+            "parts_out",
+            "annotations",
+        ]
+
+        return (ShuffleLayer, tuple(getattr(self, attr) for attr in attrs))
+
+    def __dask_distributed_pack__(self, *args, **kwargs):
+        ret = super().__dask_distributed_pack__(*args, **kwargs)
+        ret["inputs"] = self.inputs
+        ret["stage"] = self.stage
+        ret["nsplits"] = self.nsplits
+        return ret
+
+    def _cull_dependencies(self, keys, parts_out=None):
+        """Determine the necessary dependencies to produce `keys`.
+
+        Does not require graph materialization.
+        """
+        deps = defaultdict(set)
+        parts_out = parts_out or self._keys_to_parts(keys)
+        inp_part_map = {inp: i for i, inp in enumerate(self.inputs)}
+        for part in parts_out:
+            out = self.inputs[part]
+            for k in range(self.nsplits):
+                _inp = insert(out, self.stage, k)
+                _part = inp_part_map[_inp]
+                if self.stage == 0 and _part >= self.npartitions_input:
+                    deps[(self.name, part)].add(("group-" + self.name, _inp, "empty"))
+                else:
+                    deps[(self.name, part)].add((self.name_input, _part))
+        return deps
+
+    def _cull(self, parts_out):
+        return ShuffleLayer(
+            self.name,
+            self.column,
+            self.inputs,
+            self.stage,
+            self.npartitions,
+            self.npartitions_input,
+            self.nsplits,
+            self.ignore_index,
+            self.name_input,
+            self.meta_input,
+            parts_out=parts_out,
+        )
+
+    def _construct_graph(self):
         """Construct graph for a "rearrange-by-column" stage."""
 
-        shuffle_group_name = "group-" + name
-        shuffle_split_name = "split-" + name
+        shuffle_group_name = "group-" + self.name
+        shuffle_split_name = "split-" + self.name
 
         dsk = {}
-        inp_part_map = {inp: i for i, inp in enumerate(inputs)}
-        for part in parts_out:
+        inp_part_map = {inp: i for i, inp in enumerate(self.inputs)}
+        for part in self.parts_out:
 
-            out = inputs[part]
+            out = self.inputs[part]
 
             _concat_list = []  # get_item tasks to concat for this output partition
-            for i in range(nsplits):
+            for i in range(self.nsplits):
                 # Get out each individual dataframe piece from the dicts
-                _inp = insert(out, stage, i)
-                _idx = out[stage]
+                _inp = insert(out, self.stage, i)
+                _idx = out[self.stage]
                 _concat_list.append((shuffle_split_name, _idx, _inp))
 
             # concatenate those pieces together, with their friends
-            dsk[(name, part)] = (
-                run_ext_function,
-                concat_func,
+            dsk[(self.name, part)] = (
+                self.concat_func,
                 _concat_list,
-                ignore_index,
+                self.ignore_index,
             )
 
             for _, _idx, _inp in _concat_list:
                 dsk[(shuffle_split_name, _idx, _inp)] = (
-                    run_ext_function,
-                    getitem_func,
+                    operator.getitem,
                     (shuffle_group_name, _inp),
                     _idx,
                 )
@@ -167,28 +424,27 @@ class ShuffleGraph(SimpleShuffleGraph):
 
                     # Initial partitions (output of previous stage)
                     _part = inp_part_map[_inp]
-                    if stage == 0:
-                        if _part < npartitions_input:
-                            input_key = (name_input, _part)
+                    if self.stage == 0:
+                        if _part < self.npartitions_input:
+                            input_key = (self.name_input, _part)
                         else:
                             # In order to make sure that to_serialize() serialize the
                             # empty dataframe input, we add it as a key.
                             input_key = (shuffle_group_name, _inp, "empty")
-                            dsk[input_key] = meta_input
+                            dsk[input_key] = self.meta_input
                     else:
-                        input_key = (name_input, _part)
+                        input_key = (self.name_input, _part)
 
                     # Convert partition into dict of dataframe pieces
                     dsk[(shuffle_group_name, _inp)] = (
-                        run_ext_function,
-                        shuffle_group_func,
+                        self.shuffle_group_func,
                         input_key,
-                        column,
-                        stage,
-                        nsplits,
-                        npartitions_input,
-                        ignore_index,
-                        npartitions,
+                        self.column,
+                        self.stage,
+                        self.nsplits,
+                        self.npartitions_input,
+                        self.ignore_index,
+                        self.npartitions,
                     )
 
         return dsk
